fix: move external provenance checks into library
